requirements.txt
Flask
psycopg2-binary
python-dotenv
groq
streamlit
flask-cors
requests